+++ views/common/_calendarTypeItem.php
@@ -6,8 +6,10 @@
  *
  */
 
-use yii\helpers\Html;
+use humhub\modules\admin\permissions\ManageModules;
+use humhub\widgets\Label;
 use humhub\widgets\ModalButton;
+use yii\helpers\Html;
 
 /* @var $editUrl string */
 /* @var $deleteUrl string */
@@ -23,11 +25,17 @@ use humhub\widgets\ModalButton;
             <span class="input-group-addon">
                 <i style="display:inline-block;width:16px;height:16px;background-color: <?= Html::encode($color)?>"></i>
             </span>
-            <input class="form-control" value="<?= Html::encode($title) ?><?= ($disabled ? ' - '.Yii::t('CalendarModule.config', '(disabled)') : '') ?>" title="<?= Yii::t('CalendarModule.base', 'Event type color');?>" type="text" readonly>
+            <span class="form-control" readonly>
+                <?= Html::encode($title) ?>
+                <?php if ($disabled) : ?>
+                    <?= Label::warning(Yii::t('CalendarModule.base', 'disabled')) ?>
+                <?php endif; ?>
+                <?php if ($isSpaceGlobal) : ?>
+                    <?= Label::info(Yii::t('CalendarModule.base', 'global')) ?>
+                <?php endif; ?>
+            </span>
             <span class="input-group-addon">
-                <?php if($isSpaceGlobal) : ?>
-                    <small><?=  Yii::t('CalendarModule.config', '(global)') ?></small>
-                <?php else: ?>
+                <?php if(!$isSpaceGlobal || Yii::$app->user->can([ManageModules::class])) : ?>
                     <?= ModalButton::primary()->load($editUrl)->icon('fa-pencil')->xs() ?>
                     <?php if(!empty($deleteUrl)) : ?>
                         <?= ModalButton::danger()->post($deleteUrl)->confirm(