+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 -----------------------
 - Enh #424: Tests for `next` version
 - Enh #423: Add Create Button
+- Enh #379: Allow managing of global event types from container side
 - Enh #382: Allow inviting by users who can manage a calendar entry
 
 1.5.3 (September 14, 2023)